+++ src/shogun/mathematics/graph/Graph.h
@@ -4,7 +4,7 @@
 #include <deque>
 #include <memory>
 #include <shogun/mathematics/graph/GraphExecutor.h>
-#include <shogun/mathematics/graph/node_implementation/NodeImplementation.h>
+#include <shogun/mathematics/graph/runtime/RuntimeNode.h>
 #include <shogun/mathematics/graph/nodes/Input.h>
 
 #include <unordered_set>
