+++ src/shogun/mathematics/graph/ShogunGraph.cpp
@@ -1,9 +1,9 @@
 #include <shogun/mathematics/graph/ShogunGraph.h>
-#include <shogun/mathematics/graph/node_implementation/shogun/Add.h>
-#include <shogun/mathematics/graph/node_implementation/shogun/Divide.h>
-#include <shogun/mathematics/graph/node_implementation/shogun/Input.h>
-#include <shogun/mathematics/graph/node_implementation/shogun/Multiply.h>
-#include <shogun/mathematics/graph/node_implementation/shogun/Subtract.h>
+#include <shogun/mathematics/graph/runtime/shogun/Add.h>
+#include <shogun/mathematics/graph/runtime/shogun/Divide.h>
+#include <shogun/mathematics/graph/runtime/shogun/Input.h>
+#include <shogun/mathematics/graph/runtime/shogun/Multiply.h>
+#include <shogun/mathematics/graph/runtime/shogun/Subtract.h>
 #include <shogun/mathematics/graph/nodes/Node.h>
 
 using namespace shogun::graph;
