+++ src/shogun/mathematics/graph/runtime/ngraph/Divide.h
@@ -4,13 +4,13 @@
  * Authors: Gil Hoben
  */
 
-#ifndef SHOGUN_SUBTRACT_NGRAPH_H_
-#define SHOGUN_SUBTRACT_NGRAPH_H_
+#ifndef SHOGUN_DIVIDE_NGRAPH_H_
+#define SHOGUN_DIVIDE_NGRAPH_H_
 
-#include <shogun/mathematics/graph/node_implementation/NodeImplementation.h>
-#include <shogun/mathematics/graph/nodes/Subtract.h>
+#include <shogun/mathematics/graph/runtime/RuntimeNode.h>
+#include <shogun/mathematics/graph/nodes/Divide.h>
 
-#include <ngraph/op/subtract.hpp>
+#include <ngraph/op/divide.hpp>
 
 namespace shogun
 {
@@ -20,27 +20,27 @@ namespace shogun
 		{
 			namespace ngraph
 			{
-				IGNORE_IN_CLASSLIST class SubtractNGraph
+				IGNORE_IN_CLASSLIST class DivideNGraph
 
-				    : public RuntimeNodeTemplate<node::Subtract, ::ngraph::Node>
+				    : public RuntimeNodeTemplate<node::Divide, ::ngraph::Node>
 				{
 				public:
-					SubtractNGraph() : RuntimeNodeTemplate()
+					DivideNGraph() : RuntimeNodeTemplate()
 					{
 					}
 
 					std::string_view get_runtime_node_name() const final
 					{
-						return "Subtract";
+						return "Divide";
 					}
 
-					std::shared_ptr<::ngraph::Node> build_implementation(
+					[[nodiscard]] std::shared_ptr<::ngraph::Node> build_implementation(
 					    const std::shared_ptr<node::Node>& node) const final
 					{
 						if (m_input_nodes.size() != 2)
 							error("Expected two input nodes in "
-							      "SubtractNGraph.");
-						return std::make_shared<::ngraph::op::Subtract>(
+							      "DivideNGraph.");
+						return std::make_shared<::ngraph::op::Divide>(
 						    m_input_nodes[0], m_input_nodes[1]);
 					}
 				};
