+++ src/shogun/mathematics/graph/runtime/ngraph/Subtract.h
@@ -4,13 +4,13 @@
  * Authors: Gil Hoben
  */
 
-#ifndef SHOGUN_ADD_NGRAPH_H_
-#define SHOGUN_ADD_NGRAPH_H_
+#ifndef SHOGUN_SUBTRACT_NGRAPH_H_
+#define SHOGUN_SUBTRACT_NGRAPH_H_
 
-#include <shogun/mathematics/graph/node_implementation/NodeImplementation.h>
-#include <shogun/mathematics/graph/nodes/Add.h>
+#include <shogun/mathematics/graph/runtime/RuntimeNode.h>
+#include <shogun/mathematics/graph/nodes/Subtract.h>
 
-#include <ngraph/op/add.hpp>
+#include <ngraph/op/subtract.hpp>
 
 namespace shogun
 {
@@ -20,27 +20,27 @@ namespace shogun
 		{
 			namespace ngraph
 			{
-				IGNORE_IN_CLASSLIST class AddNGraph
+				IGNORE_IN_CLASSLIST class SubtractNGraph
 
-				    : public RuntimeNodeTemplate<node::Add, ::ngraph::Node>
+				    : public RuntimeNodeTemplate<node::Subtract, ::ngraph::Node>
 				{
 				public:
-					AddNGraph() : RuntimeNodeTemplate()
+					SubtractNGraph() : RuntimeNodeTemplate()
 					{
 					}
 
 					std::string_view get_runtime_node_name() const final
 					{
-						return "Add";
+						return "Subtract";
 					}
 
-					std::shared_ptr<::ngraph::Node> build_implementation(
+					[[nodiscard]] std::shared_ptr<::ngraph::Node> build_implementation(
 					    const std::shared_ptr<node::Node>& node) const final
 					{
 						if (m_input_nodes.size() != 2)
 							error("Expected two input nodes in "
-							      "AddNGraph.");
-						return std::make_shared<::ngraph::op::Add>(
+							      "SubtractNGraph.");
+						return std::make_shared<::ngraph::op::Subtract>(
 						    m_input_nodes[0], m_input_nodes[1]);
 					}
 				};
