+++ src/shogun/mathematics/graph/runtime/shogun/Divide.h
@@ -4,12 +4,12 @@
  * Authors: Gil Hoben
  */
 
-#ifndef SHOGUN_DETAIL_SUBTRACT_SHOGUN_H_
-#define SHOGUN_DETAIL_SUBTRACT_SHOGUN_H_
+#ifndef SHOGUN_DETAIL_DIVIDE_SHOGUN_H_
+#define SHOGUN_DETAIL_DIVIDE_SHOGUN_H_
 
-#include <shogun/mathematics/graph/node_implementation/shogun/ShogunBinaryNode.h>
-#include <shogun/mathematics/graph/nodes/Subtract.h>
-#include <shogun/mathematics/graph/ops/shogun/Subtract.h>
+#include <shogun/mathematics/graph/runtime/shogun/ShogunBinaryNode.h>
+#include <shogun/mathematics/graph/nodes/Divide.h>
+#include <shogun/mathematics/graph/ops/shogun/Divide.h>
 
 namespace shogun
 {
@@ -19,18 +19,18 @@ namespace shogun
 		{
 			namespace shogun
 			{
-				IGNORE_IN_CLASSLIST class SubtractShogun
+				IGNORE_IN_CLASSLIST class DivideShogun
 				    : public ShogunBinaryRuntimeNode<
-				          SubtractShogun, node::Subtract, OutputNode>
+				          DivideShogun, node::Divide, OutputNode>
 				{
 				public:
-					SubtractShogun() : ShogunBinaryRuntimeNode()
+					DivideShogun() : ShogunBinaryRuntimeNode()
 					{
 					}
 
 					std::string_view get_runtime_node_name() const final
 					{
-						return "Subtract";
+						return "Divide";
 					}
 
 					[[nodiscard]] std::shared_ptr<OutputNode>
@@ -39,7 +39,7 @@ namespace shogun
 					    const std::shared_ptr<OutputNode>& node2,
 					    const std::shared_ptr<node::Node>& graph_node) const {
 						return std::make_shared<OutputNode>(
-						    std::make_shared<op::SubtractShogun>(graph_node),
+						    std::make_shared<op::DivideShogun>(graph_node),
 						    node1, node2);
 					}
 				};
